API_talantix/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
mypy>=1.8.0
black>=24.1.0
isort>=5.13.0
pytest>=8.0.0
apache-airflow>=2.7.0 